+++ inc/search-form.php
@@ -16,13 +16,3 @@ function responsive_bu_search_form_contexts() {
 	);
 }
 add_filter( 'bu_search_form_contexts', 'responsive_bu_search_form_contexts' );
-
-/**
- * Add a placeholder attribute to the search form added by BU CMS.
- *
- * @return string Modified search field input tag attributes.
- */
-function responsive_bu_search_form_query_attributes() {
-	return sprintf( 'placeholder="%s"', esc_attr__( 'Search site&hellip;', 'responsive-framework' ) );
-}
-add_filter( 'bu_search_form_query_attributes', 'responsive_bu_search_form_query_attributes' );